run bundle gem facebook_data_analyzer
