+++ .gitignore
@@ -1,2 +1,15 @@
 *.xlsx
 example/facebook-monaleigh/messages/*.json
+
+# bundle gem
+/.bundle/
+/.yardoc
+/_yardoc/
+/coverage/
+/doc/
+/pkg/
+/spec/reports/
+/tmp/
+
+# rspec failure tracking
+.rspec_status
