+++ lib/facebook_data_analyzer/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module FacebookDataAnalyzer
+  VERSION = '0.1.0'
+end
